notification bug fixes
