+++ app/src/main/AndroidManifest.xml
@@ -11,6 +11,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
 
+
         <activity
             android:name="com.blog.app.activity.BlogPostListActivity"
             android:configChanges="keyboardHidden|orientation|screenSize"
@@ -37,12 +38,6 @@
         <service android:name="com.blog.app.service.MyFirebaseMessagingService">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
-            </intent-filter>
-        </service>
-
-        <service android:name="com.blog.app.service.MyFirebaseInstanceIDService">
-            <intent-filter>
-                <action android:name="com.google.firebase.INSTANCE_ID_EVENT" />
             </intent-filter>
         </service>
         <!-- ./Firebase Notifications -->
