+++ app/src/main/java/com/blog/app/config/AppConfig.java
@@ -4,5 +4,5 @@ package com.blog.app.config;
 public class AppConfig {
 
 
-    public static final String BLOG_BASE_URL = "http://techcrunch.com/";
+    public static final String BLOG_BASE_URL = "http://rahulkmishra20.000webhostapp.com/";
 }
